Finished server configuration updates
We got the https sever running on Thursday. But, as we all know, there is a large gap between working and finished.

I have reformatted the .conf files, deleted an unused .conf, and made (and tested) a shell script (ubuntu_server.sh) to start from a blank virtual machine and get us back to where we are right now. It is not enough to get to a point on our mountain, we must build train tracks to take us there anytime we want to go.

An important changed is that there are two versions of a number of files now, one for the server with and SSL certificate and access to port 433 for https. As well as a set for testing on personal machines, which is also used setup and obtain and SSL certificate.

Deploy the local versions with the docker-compose.yml file, using the same commands as outlined in the README.md. Server deployment is needs the additional, modifications '-f docker-compose-deploy.yml'. See update.sh and update-deploy.sh for working examples of each set of commands.