.env added, moved to `scripts` dir, one compose.yaml file
The configuration of the backend (Django-API) is now done
with environment variables. These variables have
defaults that are good for development and build time.

The outer dir was too busy with scripts, so I moved
the non-primary scripts to a `scripts` directory.

With the use of environment variables, we only need
a single docker `compose.yaml` file.
This file has everything needed for development,
build, and deploying the website.

The NGINX files (niginx/*.conf) were renamed.
They were longer and more specific then they needed
to be given the context of their directory