Using cached fetches at Docker-build time.
1. Updates needed from running on a mac cpu

2. Undoing a failed merge.

3. NGINX files changed to allow the NGINX not to fail if the frontend
service is not deployed at the same time as the banckend.

4. To use cached fetches durinf the frontend docker build,
we now do a build on the localhost computer (npm run build).
The only point of this is to use cached data as a fallback when
the fetches (expectedly) fail in docker during the container build phase.

5. Scripts are also updated to delete volumes data now that we use
django-static to serve the API static files.