Updates based on deployment tests.
absolute_redirect off; was required, this was mistakenly set to "on".

Building the frontend container was moved to the deploy.sh and display.sh
scripts from the frontend-build.sh.

update.sh had some text from echo command changed to look better over
a remote connection with only a single line of output.