Better security. API build from Alpine Python image.
The only thing that is needs to work in this API container is
Python. So healthcheck.py was created to replace a call to 'curl'.

start.sh was removed, the components in that script
moved into the existing settings.py.
Again, this is to only run systems using Python,
and to not depend on system packages.

The volumes connected to the API can now be toggled to be
read-only with the addition of the VOLUME_SPECIFICATION,
see .env.example.