Better security. API docker image now a read-only file system.
Added Parietal upgrades to MySQL-docker-image to
create and use SSL certificates.
MySQL has this available, but it is not being set as required.

Volume management in compose.yaml.
Added rm_volumes.sh to remove volumes
that should not be persistent across builds/runs.

Made (more) minimalist permissions for files in the docker container,
tested be working for the data.sh, display.sh.