`spectrum_output_filename` in the spectrum table looked like a URL.
Google kept trying to crawl the API because there is a full file path,
not a relative file path for the values of `spectrum_output_filename`.
Google thought it was a relative url. But this is also a security issue,
there was no need to show the file system of the computer that was making
the output files.