Container memory limits - Changed restart policy
A new prescriptive approach for container memory management.
Memory pressure was causing some containers to become zombies
as critical processes were killed in the container,
but **without** causing the container to exit and restart.

Containers have memory limits appropriately balanced for 4 GBs host.
Further adjustments (tuning) of the memory values may be required.

The RESTART_POLICY is no longer a configurable variable,
and has been removed from the documentation and examples.

Replacing the old restart policy is a new per-container
`deploy: restart_policy` that allows a limited number of container restarts
for conditions specific to that container/service.