Changes to the data migration process.
Needed stricter controls to only have data migrated to the
live website when absolutely intended.

The previous configuration would migrate staged data
when the API restarted.
Now we are guaranteeing that the data is staged in MySQL
and that TXT and FITs files have been uploaded
and the docker containers have been updated
before data can be flagged for migration.